LPS-77842 Refactor into UpgradeProcess for 7.0.x (fix needs to be backported to 7.0.x as well) (update for LPS-77842)
