+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_6.java
@@ -16,6 +16,7 @@ package com.liferay.portal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
+import com.liferay.portal.upgrade.v7_0_6.UpgradeRepository;
 import com.liferay.portal.upgrade.v7_0_6.UpgradeThemeId;
 
 /**
@@ -30,6 +31,7 @@ public class UpgradeProcess_7_0_6 extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
+		upgrade(new UpgradeRepository());
 		upgrade(new UpgradeThemeId());
 
 		clearIndexesCache();
