+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_6/util/RepositoryTable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_1_x.util;
+package com.liferay.portal.upgrade.v7_0_6.util;
 
 import java.sql.Types;
